+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/langchain
 
+## 1.0.0-beta.28
+
+### Patch Changes
+
+- Updated dependencies [84343eb]
+- Updated dependencies [a753b3a]
+  - ai@5.0.0-beta.28
+
+## 1.0.0-beta.27
+
+### Patch Changes
+
+- Updated dependencies [d5ae088]
+  - ai@5.0.0-beta.27
+
 ## 1.0.0-beta.26
 
 ### Patch Changes