+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### Patch Changes
 
+- Updated dependencies [84343eb]
+- Updated dependencies [a753b3a]
+  - ai@5.0.0-beta.28
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [d5ae088]
+  - ai@5.0.0-beta.27
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [ae77a99]
 - Updated dependencies [1f6ce57]
 - Updated dependencies [5a975a4]