+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ai-sdk/svelte
 
+## 3.0.0-beta.28
+
+### Patch Changes
+
+- ac34802: Add clear object function to StructuredObject
+- Updated dependencies [ac34802]
+- Updated dependencies [84343eb]
+- Updated dependencies [a753b3a]
+  - @ai-sdk/provider-utils@3.0.0-beta.6
+  - ai@5.0.0-beta.28
+
+## 3.0.0-beta.27
+
+### Patch Changes
+
+- Updated dependencies [d5ae088]
+  - ai@5.0.0-beta.27
+
 ## 3.0.0-beta.26
 
 ### Patch Changes